% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.did_multiplegt_stat}
\alias{print.did_multiplegt_stat}
\title{print method for did_multiplegt_stat}
\usage{
\method{print}{did_multiplegt_stat}(x, ...)
}
\arguments{
\item{x}{A did_multiplegt_stat object}

\item{...}{Undocumented}
}
\value{
No return, just a custom summary print for did_multiplegt_stat output.
}
\description{
A customized printed display for did_continous output
}
