% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.did_continuous}
\alias{summary.did_continuous}
\title{summary method for did_continuous}
\usage{
\method{summary}{did_continuous}(object, ...)
}
\arguments{
\item{object}{A did_continuous object}

\item{...}{Undocumented}
}
\description{
A customized printed display for did_continuous output
}
