% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.did_multiplegt_stat}
\alias{summary.did_multiplegt_stat}
\title{summary method for did_multiplegt_stat}
\usage{
\method{summary}{did_multiplegt_stat}(object, ...)
}
\arguments{
\item{object}{A did_multiplegt_stat object}

\item{...}{Undocumented}
}
\value{
No return, just a custom summary method for did_multiplegt_stat output.
}
\description{
A customized printed display for did_multiplegt_stat output
}
