% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnames.R
\name{rnames.did_multiplegt_stat}
\alias{rnames.did_multiplegt_stat}
\title{rnames method for did_multiplegt_stat}
\usage{
rnames.did_multiplegt_stat(
  object,
  ignore = c("by_fd_graph", "by_graph", "args"),
  ...
)
}
\arguments{
\item{object}{A did_multiplegt_stat object}

\item{ignore}{Ignored sublists}

\item{...}{Undocumented}
}
\value{
The same output as rnames.
}
\description{
A customized rnames method for did_multiplegt_stat output
}
